+++ emborg/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.20.0"
-__released__ = "2021-02-13"
+__version__ = "1.20.1"
+__released__ = "2021-03-02"
 
 from .settings import Settings as Emborg